refactor(crane-context): split notifications.ts into write/query/status/aggregate
Brings workers/crane-context/src/notifications.ts under structural caps
(max-lines: 500, max-lines-per-function: 75, complexity: 15).

- notifications.ts (orchestrator, 447 LOC, was 1114): repoToVenture,
  computeDedupeHash, buildMatchKey, resolveNotificationsByBranch,
  runStaleBranchSweep, runNonMainCleanup, processGreenEvent + re-exports
- notifications-write.ts: createNotification (complexity reduced via
  extracted buildInsertBinds + n() null-coerce helper)
- notifications-query.ts: listNotifications
- notifications-status.ts: updateNotificationStatus + VALID_TRANSITIONS
- notifications-aggregate.ts: countNotifications, countUnresolved,
  getOldestNotification (split into buildFilterConditions, tallySeverityRows,
  tallyStatusRows, tallyVentureRows helpers)

Public API unchanged. 540/540 tests pass.
