+++ workers/crane-context/src/notifications.ts
@@ -5,30 +5,27 @@
  * Handles deduplication, retention filtering, and venture derivation.
  */
 
-import {
-  VENTURE_CONFIG,
-  NOTIFICATION_RETENTION_DAYS,
-  NOTIFICATION_STATUSES,
-  MAX_NOTIFICATION_DETAILS_SIZE,
-  DEFAULT_PAGE_SIZE,
-  MAX_PAGE_SIZE,
-} from './constants'
-import type {
-  NotificationRecord,
-  NotificationSeverity,
-  NotificationStatus,
-  NotificationMatchKeyVersion,
-  NotificationAutoResolveReason,
-} from './types'
-import {
-  generateNotificationId,
-  nowIso,
-  sha256,
-  sizeInBytes,
-  encodeCursor,
-  decodeCursor,
-} from './utils'
+import { VENTURE_CONFIG, NOTIFICATION_RETENTION_DAYS } from './constants'
+import type { NotificationMatchKeyVersion, NotificationAutoResolveReason } from './types'
+import { nowIso, sha256, sizeInBytes, generateNotificationId } from './utils'
 import { logNotificationEvent } from './notifications-log'
+import { MAX_NOTIFICATION_DETAILS_SIZE } from './constants'
+
+// Re-export from split modules (public API unchanged)
+export type { CreateNotificationParams, CreateNotificationResult } from './notifications-write'
+export { createNotification } from './notifications-write'
+
+export type { ListNotificationsParams, ListNotificationsResult } from './notifications-query'
+export { listNotifications } from './notifications-query'
+
+export { updateNotificationStatus } from './notifications-status'
+
+export type { CountNotificationsParams, NotificationCountsResult } from './notifications-aggregate'
+export {
+  countUnresolved,
+  countNotifications,
+  getOldestNotification,
+} from './notifications-aggregate'
 
 // ============================================================================
 // Venture Derivation
@@ -66,8 +63,8 @@ export async function computeDedupeHash(params: {
   const input = [
     params.source,
     params.event_type,
-    params.repo || '',
-    params.branch || '',
+    params.repo ?? '',
+    params.branch ?? '',
     params.content_key,
   ].join('|')
   return sha256(input)
@@ -77,21 +74,14 @@ export async function computeDedupeHash(params: {
 // Match Key Construction
 // ============================================================================
 
-/**
- * Inputs needed to compute a v2_id (numeric workflow_id) match key for a
- * GitHub workflow_run event.
- */
 export interface BuildWorkflowRunMatchKeyParams {
   source: 'github'
   kind: 'workflow_run'
-  repo_full_name: string // owner/repo - never bare repo name
+  repo_full_name: string
   branch: string
   workflow_id: number
 }
 
-/**
- * Inputs needed to compute a v2_id match key for a GitHub check_suite event.
- */
 export interface BuildCheckSuiteMatchKeyParams {
   source: 'github'
   kind: 'check_suite'
@@ -100,9 +90,6 @@ export interface BuildCheckSuiteMatchKeyParams {
   app_id: number
 }
 
-/**
- * Inputs needed to compute a v2_id match key for a GitHub check_run event.
- */
 export interface BuildCheckRunMatchKeyParams {
   source: 'github'
   kind: 'check_run'
@@ -116,15 +103,9 @@ export interface BuildCheckRunMatchKeyParams {
  * Inputs needed to compute a match key for a Vercel deployment event.
  *
  * `vercel_team_id` is REQUIRED for cross-team collision safety. Two Vercel
- * projects in different teams can have the same `project_name` (operators
- * routinely call sites "marketing-site" or "console"). Without team_id in
- * the match key, a green deployment in one team would silently auto-resolve
- * a red deployment in another - the same class of cross-org collision bug
- * the GitHub match keys protect against via `owner/repo`.
- *
- * For Vercel events that arrive without a team_id (legacy webhook payloads
- * or single-team setups), pass the literal string "no-team" so the slot
- * is still occupied and the format remains unambiguous.
+ * projects in different teams can have the same `project_name`. Without
+ * team_id, a green deployment in one team would silently auto-resolve a red
+ * deployment in another. Pass "no-team" for legacy payloads without team_id.
  */
 export interface BuildVercelMatchKeyParams {
   source: 'vercel'
@@ -143,16 +124,7 @@ export type BuildMatchKeyParams =
 
 /**
  * Construct the canonical v2_id match key for a notification.
- *
- * The match key is the unique identifier the auto-resolver uses to find
- * prior failure notifications matching a green event. It uses the FULL
- * owner/repo (never the bare repo name) to prevent cross-org collisions:
- * a green from venturecrane/console must NOT auto-resolve a red from
- * siliconcrane/console.
- *
- * Returns the match key string and version marker. Version is always
- * 'v2_id' for new code; the legacy 'v1_name' format is only produced by
- * the migration 0023 backfill.
+ * Uses full owner/repo to prevent cross-org collisions.
  */
 export function buildMatchKey(params: BuildMatchKeyParams): {
   match_key: string
@@ -171,13 +143,11 @@ export function buildMatchKey(params: BuildMatchKeyParams): {
         match_key_version: 'v2_id',
       }
     }
-    // check_run
     return {
       match_key: `gh:cr:${params.repo_full_name}:${params.branch}:${params.app_id}:${params.name}`,
       match_key_version: 'v2_id',
     }
   }
-  // vercel - includes vercel_team_id to prevent cross-team collision
   return {
     match_key: `vc:dpl:${params.repo_full_name}:${params.branch}:${params.vercel_team_id}:${params.project_name}:${params.target}`,
     match_key_version: 'v2_id',
@@ -185,313 +155,8 @@ export function buildMatchKey(params: BuildMatchKeyParams): {
 }
 
 // ============================================================================
-// Create
-// ============================================================================
-
-export interface CreateNotificationParams {
-  source: string
-  event_type: string
-  severity: NotificationSeverity
-  summary: string
-  details_json: string
-  external_id?: string
-  dedupe_hash: string
-  venture?: string | null
-  repo?: string | null
-  branch?: string | null
-  environment?: string | null
-  created_at?: string
-  actor_key_id: string
-
-  // Match-key fields (PR A2). Optional for backward compatibility with
-  // call sites that haven't been updated yet, but new failure normalizers
-  // populate them so subsequent green events can match.
-  workflow_id?: number | null
-  workflow_name?: string | null
-  run_id?: number | null
-  head_sha?: string | null
-  check_suite_id?: number | null
-  check_run_id?: number | null
-  app_id?: number | null
-  app_name?: string | null
-  deployment_id?: string | null
-  project_name?: string | null
-  target?: string | null
-  match_key?: string | null
-  match_key_version?: NotificationMatchKeyVersion | null
-  run_started_at?: string | null
-}
-
-export interface CreateNotificationResult {
-  notification?: NotificationRecord
-  duplicate: boolean
-}
-
-/**
- * Insert a notification, silently ignoring duplicates (INSERT OR IGNORE on dedupe_hash).
- */
-export async function createNotification(
-  db: D1Database,
-  params: CreateNotificationParams
-): Promise<CreateNotificationResult> {
-  // Validate details_json size
-  if (sizeInBytes(params.details_json) > MAX_NOTIFICATION_DETAILS_SIZE) {
-    throw new Error(`details_json exceeds maximum size of ${MAX_NOTIFICATION_DETAILS_SIZE} bytes`)
-  }
-
-  const id = generateNotificationId()
-  const now = nowIso()
-  const createdAt = params.created_at || now
-
-  const result = await db
-    .prepare(
-      `INSERT OR IGNORE INTO notifications
-       (id, source, event_type, severity, status, summary, details_json,
-        external_id, dedupe_hash, venture, repo, branch, environment,
-        created_at, received_at, updated_at, actor_key_id,
-        workflow_id, workflow_name, run_id, head_sha,
-        check_suite_id, check_run_id, app_id, app_name,
-        deployment_id, project_name, target,
-        match_key, match_key_version, run_started_at)
-       VALUES (?, ?, ?, ?, 'new', ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,
-               ?, ?, ?, ?,
-               ?, ?, ?, ?,
-               ?, ?, ?,
-               ?, ?, ?)`
-    )
-    .bind(
-      id,
-      params.source,
-      params.event_type,
-      params.severity,
-      params.summary,
-      params.details_json,
-      params.external_id || null,
-      params.dedupe_hash,
-      params.venture || null,
-      params.repo || null,
-      params.branch || null,
-      params.environment || null,
-      createdAt,
-      now,
-      now,
-      params.actor_key_id,
-      params.workflow_id ?? null,
-      params.workflow_name ?? null,
-      params.run_id ?? null,
-      params.head_sha ?? null,
-      params.check_suite_id ?? null,
-      params.check_run_id ?? null,
-      params.app_id ?? null,
-      params.app_name ?? null,
-      params.deployment_id ?? null,
-      params.project_name ?? null,
-      params.target ?? null,
-      params.match_key ?? null,
-      params.match_key_version ?? null,
-      params.run_started_at ?? null
-    )
-    .run()
-
-  // If no rows changed, it was a duplicate (dedupe_hash already existed)
-  if (result.meta.changes === 0) {
-    return { duplicate: true }
-  }
-
-  // Fetch the created record
-  const record = await db
-    .prepare('SELECT * FROM notifications WHERE id = ?')
-    .bind(id)
-    .first<NotificationRecord>()
-
-  if (record) {
-    logNotificationEvent('notification_created', {
-      id: record.id,
-      source: record.source,
-      severity: record.severity,
-      match_key: record.match_key,
-      repo: record.repo,
-      branch: record.branch,
-      workflow_id: record.workflow_id,
-      dedupe_hash: record.dedupe_hash.slice(0, 8),
-    })
-  }
-
-  return { notification: record || undefined, duplicate: false }
-}
-
-// ============================================================================
-// List / Query
-// ============================================================================
-
-export interface ListNotificationsParams {
-  status?: string
-  severity?: string
-  venture?: string
-  repo?: string
-  source?: string
-  limit?: number
-  cursor?: string
-}
-
-export interface ListNotificationsResult {
-  notifications: NotificationRecord[]
-  next_cursor?: string
-}
-
-/**
- * List notifications with filtering and pagination.
- * Automatically applies 30-day retention filter.
- */
-export async function listNotifications(
-  db: D1Database,
-  params: ListNotificationsParams
-): Promise<ListNotificationsResult> {
-  const limit = Math.min(params.limit || DEFAULT_PAGE_SIZE, MAX_PAGE_SIZE)
-  const retentionCutoff = new Date(
-    Date.now() - NOTIFICATION_RETENTION_DAYS * 24 * 60 * 60 * 1000
-  ).toISOString()
-
-  const conditions: string[] = ['created_at > ?']
-  const binds: (string | number)[] = [retentionCutoff]
-
-  if (params.status) {
-    conditions.push('status = ?')
-    binds.push(params.status)
-  }
-  if (params.severity) {
-    conditions.push('severity = ?')
-    binds.push(params.severity)
-  }
-  if (params.venture) {
-    conditions.push('venture = ?')
-    binds.push(params.venture)
-  }
-  if (params.repo) {
-    conditions.push('repo = ?')
-    binds.push(params.repo)
-  }
-  if (params.source) {
-    conditions.push('source = ?')
-    binds.push(params.source)
-  }
-
-  if (params.cursor) {
-    const decoded = decodeCursor(params.cursor)
-    conditions.push('(created_at < ? OR (created_at = ? AND id < ?))')
-    binds.push(decoded.timestamp, decoded.timestamp, decoded.id)
-  }
-
-  const where = conditions.length > 0 ? `WHERE ${conditions.join(' AND ')}` : ''
-
-  const sql = `SELECT * FROM notifications ${where} ORDER BY created_at DESC, id DESC LIMIT ?`
-  binds.push(limit + 1)
-
-  const result = await db
-    .prepare(sql)
-    .bind(...binds)
-    .all<NotificationRecord>()
-
-  const notifications = result.results || []
-  let next_cursor: string | undefined
-
-  if (notifications.length > limit) {
-    notifications.pop()
-    const last = notifications[notifications.length - 1]
-    next_cursor = encodeCursor({ timestamp: last.created_at, id: last.id })
-  }
-
-  return { notifications, next_cursor }
-}
-
-// ============================================================================
-// Update Status
-// ============================================================================
-
-const VALID_TRANSITIONS: Record<string, string[]> = {
-  new: ['acked', 'resolved'],
-  acked: ['resolved'],
-  resolved: [],
-}
-
-/**
- * Update a notification's status with validated state transitions.
- *
- * When transitioning to 'resolved', also stamps `resolved_at` and sets
- * `auto_resolve_reason = 'manual'` so the audit trail distinguishes
- * operator-initiated resolutions from auto-resolves.
- */
-export async function updateNotificationStatus(
-  db: D1Database,
-  id: string,
-  newStatus: NotificationStatus
-): Promise<NotificationRecord | null> {
-  // Validate status value
-  if (!NOTIFICATION_STATUSES.includes(newStatus)) {
-    throw new Error(`Invalid status: ${newStatus}`)
-  }
-
-  // Get current record
-  const current = await db
-    .prepare('SELECT * FROM notifications WHERE id = ?')
-    .bind(id)
-    .first<NotificationRecord>()
-
-  if (!current) return null
-
-  // Validate state transition
-  const allowed = VALID_TRANSITIONS[current.status]
-  if (!allowed || !allowed.includes(newStatus)) {
-    throw new Error(
-      `Invalid state transition: ${current.status} -> ${newStatus}. Allowed: ${(allowed || []).join(', ') || 'none'}`
-    )
-  }
-
-  const now = nowIso()
-  if (newStatus === 'resolved') {
-    await db
-      .prepare(
-        `UPDATE notifications
-         SET status = ?, updated_at = ?, resolved_at = ?, auto_resolve_reason = 'manual'
-         WHERE id = ?`
-      )
-      .bind(newStatus, now, now, id)
-      .run()
-    logNotificationEvent('notification_resolved_manual', {
-      id,
-      prior_status: current.status,
-    })
-    return {
-      ...current,
-      status: newStatus,
-      updated_at: now,
-      resolved_at: now,
-      auto_resolve_reason: 'manual',
-    }
-  }
-
-  await db
-    .prepare('UPDATE notifications SET status = ?, updated_at = ? WHERE id = ?')
-    .bind(newStatus, now, id)
-    .run()
-
-  return { ...current, status: newStatus, updated_at: now }
-}
-
-// ============================================================================
 // Auto-Resolve (branch-deleted and stale-branch TTL)
 // ============================================================================
-//
-// Two bulk-resolve paths that complement `processGreenEvent`:
-//
-// - `resolveNotificationsByBranch` fires on GitHub `delete` webhooks with
-//   ref_type=branch. A deleted branch cannot produce a subsequent green
-//   event, so any open notifications on it are permanently stuck. Issue #563.
-//
-// - `runStaleBranchSweep` is the cron backstop for branches abandoned
-//   without an explicit delete event (force-pushed refs, detached-HEAD
-//   experiments, dropped forks). Only touches non-default branches to keep
-//   main red-signal pristine.
 
 export interface ResolveByBranchResult {
   resolved_count: number
@@ -499,8 +164,7 @@ export interface ResolveByBranchResult {
 }
 
 /**
- * Resolve every open notification on (repo, branch). Idempotent: running it
- * twice resolves nothing the second time.
+ * Resolve every open notification on (repo, branch). Idempotent.
  */
 export async function resolveNotificationsByBranch(
   db: D1Database,
@@ -509,19 +173,13 @@ export async function resolveNotificationsByBranch(
   reason: NotificationAutoResolveReason
 ): Promise<ResolveByBranchResult> {
   const now = nowIso()
-
   const openRows = await db
-    .prepare(
-      `SELECT id FROM notifications
-       WHERE status = 'new' AND repo = ? AND branch = ?`
-    )
+    .prepare(`SELECT id FROM notifications WHERE status = 'new' AND repo = ? AND branch = ?`)
     .bind(repo, branch)
     .all<{ id: string }>()
 
-  const matched_ids = (openRows.results || []).map((r) => r.id)
-  if (matched_ids.length === 0) {
-    return { resolved_count: 0, matched_ids: [] }
-  }
+  const matched_ids = (openRows.results ?? []).map((r) => r.id)
+  if (matched_ids.length === 0) return { resolved_count: 0, matched_ids: [] }
 
   await db
     .prepare(
@@ -538,7 +196,6 @@ export async function resolveNotificationsByBranch(
     reason,
     count: matched_ids.length,
   })
-
   return { resolved_count: matched_ids.length, matched_ids }
 }
 
@@ -548,18 +205,9 @@ export interface StaleBranchSweepResult {
 }
 
 /**
- * Regression alarm: resolve any open notifications on non-protected branches
- * older than N days. Default 1 day.
- *
- * The protected-branch gate at the normalizer/green-classifier layer means
- * non-protected-branch rows should never reach the DB in steady state. This
- * sweep stays as a backstop: if it ever resolves > 0 rows, something
- * bypassed the gate (a code path missed the import, a manual ingest, a
- * future refactor regressed the behavior). The warn-level log makes that
- * regression loud rather than silent.
- *
- * Never touches `main`/`master`/`production` rows — those represent real
- * red signal that deserves a human decision.
+ * Resolve open notifications on non-protected branches older than N days.
+ * Steady state should resolve zero rows; nonzero means something bypassed
+ * the protected-branch gate at ingest time.
  */
 export async function runStaleBranchSweep(
   db: D1Database,
@@ -567,7 +215,6 @@ export async function runStaleBranchSweep(
 ): Promise<StaleBranchSweepResult> {
   const cutoffIso = new Date(Date.now() - cutoffDays * 24 * 60 * 60 * 1000).toISOString()
   const now = nowIso()
-
   const result = await db
     .prepare(
       `UPDATE notifications
@@ -582,29 +229,22 @@ export async function runStaleBranchSweep(
     .run()
 
   const resolved_count = result.meta?.changes ?? 0
-
   if (resolved_count > 0) {
-    // Steady state should be zero. Nonzero = something bypassed the
-    // protected-branch gate at ingest time.
     logNotificationEvent('notifications_stale_branch_sweep_unexpected', {
       cutoff_days: cutoffDays,
       count: resolved_count,
       note: 'expected zero rows in steady state post-protected-branch-gate',
     })
   }
-
   return { resolved_count, cutoff_days: cutoffDays }
 }
 
 /**
- * One-shot variant of `runStaleBranchSweep` without the age cutoff. Used
- * post-deploy of the protected-branch gate to drain rows that were ingested
- * before the policy took effect. Idempotent; re-running once rows are
- * resolved returns count=0.
+ * One-shot drain: resolve all open non-main-branch notifications with no
+ * age cutoff. Used post-deploy of the protected-branch gate. Idempotent.
  */
 export async function runNonMainCleanup(db: D1Database): Promise<StaleBranchSweepResult> {
   const now = nowIso()
-
   const result = await db
     .prepare(
       `UPDATE notifications
@@ -618,11 +258,7 @@ export async function runNonMainCleanup(db: D1Database): Promise<StaleBranchSwee
     .run()
 
   const resolved_count = result.meta?.changes ?? 0
-
-  logNotificationEvent('notifications_non_main_cleanup', {
-    count: resolved_count,
-  })
-
+  logNotificationEvent('notifications_non_main_cleanup', { count: resolved_count })
   return { resolved_count, cutoff_days: 0 }
 }
 
@@ -645,7 +281,6 @@ export interface ProcessGreenEventParams {
   summary: string
   dedupe_hash: string
   auto_resolve_reason: NotificationAutoResolveReason
-  // Structural identifiers (carried through to the synthetic green row)
   workflow_id?: number | null
   workflow_name?: string | null
   run_id?: number | null
@@ -663,214 +298,140 @@ export interface ProcessGreenEventResult {
   green_notification_id: string | null
   resolved_count: number
   matched_ids: string[]
-  duplicate: boolean // true if this exact green was already processed
+  duplicate: boolean
+}
+
+// Normalize undefined → null for D1 bind parameters
+function n<T>(v: T | null | undefined): T | null {
+  return v ?? null
+}
+
+const GREEN_INSERT_SQL = `INSERT OR IGNORE INTO notifications
+   (id, source, event_type, severity, status, summary, details_json,
+    external_id, dedupe_hash, venture, repo, branch, environment,
+    created_at, received_at, updated_at, actor_key_id,
+    workflow_id, workflow_name, run_id, head_sha,
+    check_suite_id, check_run_id, app_id, app_name,
+    deployment_id, project_name, target,
+    match_key, match_key_version, run_started_at,
+    auto_resolve_reason, resolved_at)
+   VALUES (?, ?, ?, 'info', 'resolved', ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,
+           ?, ?, ?, ?,
+           ?, ?, ?, ?,
+           ?, ?, ?,
+           ?, ?, ?,
+           ?, ?)`
+
+function greenInsertBinds(
+  greenId: string,
+  now: string,
+  p: ProcessGreenEventParams
+): (string | number | null)[] {
+  return [
+    greenId,
+    p.source,
+    p.event_type,
+    p.summary,
+    p.details_json,
+    null,
+    p.dedupe_hash,
+    p.venture,
+    p.repo,
+    p.branch,
+    null,
+    p.run_started_at,
+    now,
+    now,
+    p.actor_key_id,
+    n(p.workflow_id),
+    n(p.workflow_name),
+    n(p.run_id),
+    p.head_sha,
+    n(p.check_suite_id),
+    n(p.check_run_id),
+    n(p.app_id),
+    n(p.app_name),
+    n(p.deployment_id),
+    n(p.project_name),
+    n(p.target),
+    p.match_key,
+    p.match_key_version,
+    p.run_started_at,
+    p.auto_resolve_reason,
+    now,
+  ]
+}
+
+async function fetchResolvedIds(
+  db: D1Database,
+  matchKey: string,
+  greenId: string
+): Promise<string[]> {
+  const rows = await db
+    .prepare(`SELECT id FROM notifications WHERE match_key = ? AND auto_resolved_by_id = ?`)
+    .bind(matchKey, greenId)
+    .all<{ id: string }>()
+  return (rows.results ?? []).map((r) => r.id)
 }
 
 /**
  * Process a green CI/CD event: insert a synthetic resolved notification row,
  * then atomically resolve all matching open prior failure notifications.
  *
- * Concurrency model (race-safe under any D1 isolation level):
- *
- *   1. INSERT OR IGNORE the green row first. The dedupe_hash UNIQUE
- *      constraint makes this idempotent: a duplicate webhook delivery
- *      results in the second insert being a no-op.
- *
- *   2. UPDATE all open notifications matching this match_key, with the
- *      idempotent predicate `WHERE auto_resolved_by_id IS NULL`. Two
- *      concurrent greens for the same match_key (different run_ids) both
- *      INSERT successfully (different dedupe_hashes), but only the first
- *      UPDATE acquires the rows by setting `auto_resolved_by_id`. The
- *      second UPDATE finds zero matching rows (the predicate fails). No
- *      double-resolution. No corrupted history.
- *
- *   3. Forward-in-time predicate: only resolve notifications whose
- *      `run_started_at` is older than (or NULL, for legacy rows) the
- *      green's `run_started_at`. Handles out-of-order webhook delivery.
- *
- *   4. Schedule-like events (cron, repository_dispatch) require same SHA:
- *      a nightly cron success the day after a nightly cron failure does
- *      NOT prove the underlying issue was fixed. Only re-running the same
- *      commit and getting green proves it.
- *
- *   5. Retention guard: never resolve notifications older than the
- *      retention window (30 days). They will fall out of the read filter
- *      anyway and there is no value in updating them.
- *
- * Returns the green notification id (if successfully inserted), the count
- * and ids of resolved prior notifications, and a `duplicate` flag if the
- * exact green event was already processed.
+ * Race safety: the `auto_resolved_by_id IS NULL` predicate in the UPDATE
+ * ensures concurrent greens for the same match_key cannot double-resolve.
+ * Forward-in-time predicate handles out-of-order webhook delivery.
+ * Schedule-like events require same head_sha.
  */
 export async function processGreenEvent(
   db: D1Database,
   params: ProcessGreenEventParams
 ): Promise<ProcessGreenEventResult> {
-  // Validate details size
   if (sizeInBytes(params.details_json) > MAX_NOTIFICATION_DETAILS_SIZE) {
     throw new Error(`details_json exceeds maximum size of ${MAX_NOTIFICATION_DETAILS_SIZE} bytes`)
   }
 
   const greenId = generateNotificationId()
   const now = nowIso()
-  const retentionCutoff = new Date(
+  const cutoff = new Date(
     Date.now() - NOTIFICATION_RETENTION_DAYS * 24 * 60 * 60 * 1000
   ).toISOString()
 
-  // Step 1: INSERT the green notification row (idempotent via dedupe_hash UNIQUE).
   const insertResult = await db
-    .prepare(
-      `INSERT OR IGNORE INTO notifications
-       (id, source, event_type, severity, status, summary, details_json,
-        external_id, dedupe_hash, venture, repo, branch, environment,
-        created_at, received_at, updated_at, actor_key_id,
-        workflow_id, workflow_name, run_id, head_sha,
-        check_suite_id, check_run_id, app_id, app_name,
-        deployment_id, project_name, target,
-        match_key, match_key_version, run_started_at,
-        auto_resolve_reason, resolved_at)
-       VALUES (?, ?, ?, 'info', 'resolved', ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,
-               ?, ?, ?, ?,
-               ?, ?, ?, ?,
-               ?, ?, ?,
-               ?, ?, ?,
-               ?, ?)`
-    )
-    .bind(
-      greenId,
-      params.source,
-      params.event_type,
-      params.summary,
-      params.details_json,
-      null, // external_id
-      params.dedupe_hash,
-      params.venture,
-      params.repo,
-      params.branch,
-      null, // environment - greens are not env-specific in the same way as failures
-      params.run_started_at,
-      now,
-      now,
-      params.actor_key_id,
-      params.workflow_id ?? null,
-      params.workflow_name ?? null,
-      params.run_id ?? null,
-      params.head_sha,
-      params.check_suite_id ?? null,
-      params.check_run_id ?? null,
-      params.app_id ?? null,
-      params.app_name ?? null,
-      params.deployment_id ?? null,
-      params.project_name ?? null,
-      params.target ?? null,
-      params.match_key,
-      params.match_key_version,
-      params.run_started_at,
-      params.auto_resolve_reason,
-      now
-    )
+    .prepare(GREEN_INSERT_SQL)
+    .bind(...greenInsertBinds(greenId, now, params))
     .run()
 
   if (insertResult.meta.changes === 0) {
-    // Duplicate green delivery. Don't double-process.
     logNotificationEvent('green_event_idempotent_skip', {
       match_key: params.match_key,
-      run_id: params.run_id ?? null,
+      run_id: n(params.run_id),
       dedupe_hash: params.dedupe_hash.slice(0, 8),
     })
-    return {
-      green_notification_id: null,
-      resolved_count: 0,
-      matched_ids: [],
-      duplicate: true,
-    }
+    return { green_notification_id: null, resolved_count: 0, matched_ids: [], duplicate: true }
   }
 
-  // Step 2: UPDATE matching open notifications. Idempotent via the
-  // `auto_resolved_by_id IS NULL` predicate (race-safe across concurrent
-  // greens for the same match_key).
-  let updateSql: string
-  let updateBinds: (string | number | null)[]
-
-  if (params.is_schedule_like) {
-    // Schedule-like events require same head_sha.
-    updateSql = `UPDATE notifications
-       SET status = 'resolved',
-           auto_resolved_by_id = ?,
-           auto_resolve_reason = ?,
-           resolved_at = ?,
-           updated_at = ?
-       WHERE match_key = ?
-         AND status IN ('new', 'acked')
-         AND auto_resolved_by_id IS NULL
-         AND created_at > ?
-         AND head_sha = ?
-         AND (run_started_at IS NULL OR run_started_at <= ?)`
-    updateBinds = [
-      greenId,
-      params.auto_resolve_reason,
-      now,
-      now,
-      params.match_key,
-      retentionCutoff,
-      params.head_sha,
-      params.run_started_at,
-    ]
-  } else {
-    // Normal events: forward-in-time predicate, any SHA on the same branch.
-    updateSql = `UPDATE notifications
-       SET status = 'resolved',
-           auto_resolved_by_id = ?,
-           auto_resolve_reason = ?,
-           resolved_at = ?,
-           updated_at = ?
-       WHERE match_key = ?
-         AND status IN ('new', 'acked')
-         AND auto_resolved_by_id IS NULL
-         AND created_at > ?
-         AND (run_started_at IS NULL OR run_started_at <= ?)`
-    updateBinds = [
-      greenId,
-      params.auto_resolve_reason,
-      now,
-      now,
-      params.match_key,
-      retentionCutoff,
-      params.run_started_at,
-    ]
-  }
-
+  const { sql, binds } = buildResolveQuery(greenId, now, cutoff, params)
   const updateResult = await db
-    .prepare(updateSql)
-    .bind(...updateBinds)
+    .prepare(sql)
+    .bind(...binds)
     .run()
+  const resolvedCount = updateResult.meta.changes ?? 0
 
-  const resolvedCount = updateResult.meta.changes || 0
-
-  // Fetch the matched ids for the audit trail (small set, this query is fine).
-  let matchedIds: string[] = []
-  if (resolvedCount > 0) {
-    const matchedRows = await db
-      .prepare(
-        `SELECT id FROM notifications
-         WHERE match_key = ? AND auto_resolved_by_id = ?`
-      )
-      .bind(params.match_key, greenId)
-      .all<{ id: string }>()
-    matchedIds = (matchedRows.results || []).map((r) => r.id)
-  }
+  const matchedIds = resolvedCount > 0 ? await fetchResolvedIds(db, params.match_key, greenId) : []
 
   if (resolvedCount > 0) {
     logNotificationEvent('success_event_received_match', {
       match_key: params.match_key,
       resolved_count: resolvedCount,
       matched_ids: matchedIds,
-      run_id: params.run_id ?? null,
+      run_id: n(params.run_id),
       head_sha: params.head_sha,
     })
   } else {
     logNotificationEvent('success_event_received_no_match', {
       match_key: params.match_key,
-      run_id: params.run_id ?? null,
+      run_id: n(params.run_id),
       head_sha: params.head_sha,
       reason: 'no_open_for_key_or_race_lost',
     })
@@ -884,231 +445,44 @@ export async function processGreenEvent(
   }
 }
 
-// ============================================================================
-// Aggregates
-// ============================================================================
-
-/**
- * Count unresolved (new + acked) notifications, optionally filtered by venture.
- */
-export async function countUnresolved(
-  db: D1Database,
-  venture?: string
-): Promise<{ total: number; critical: number; warning: number }> {
-  const retentionCutoff = new Date(
-    Date.now() - NOTIFICATION_RETENTION_DAYS * 24 * 60 * 60 * 1000
-  ).toISOString()
-
-  let sql = `SELECT severity, COUNT(*) as count FROM notifications
-    WHERE status IN ('new', 'acked') AND created_at > ?`
-  const binds: string[] = [retentionCutoff]
-
-  if (venture) {
-    sql += ' AND venture = ?'
-    binds.push(venture)
-  }
-
-  sql += ' GROUP BY severity'
-
-  const result = await db
-    .prepare(sql)
-    .bind(...binds)
-    .all<{ severity: string; count: number }>()
-
-  const counts = { total: 0, critical: 0, warning: 0 }
-  for (const row of result.results || []) {
-    counts.total += row.count
-    if (row.severity === 'critical') counts.critical = row.count
-    if (row.severity === 'warning') counts.warning = row.count
-  }
-
-  return counts
-}
-
-// ============================================================================
-// Truthful counts (Track B PR B-1)
-// ============================================================================
-
-/**
- * Truthful count of notifications matching a filter, broken down by status
- * and severity. The SOS uses this to display "270 alerts (12 critical, 45 warning)"
- * instead of `${notifications.length}` from a paginated slice.
- *
- * Plan §B.3: this is the missing endpoint that fixes the loudest defect
- * (defect #1 — SOS displaying "10 unresolved" when DB has 270).
- */
-export interface CountNotificationsParams {
-  status?: string
-  severity?: string
-  venture?: string
-  repo?: string
-  source?: string
-  /**
-   * When 'venture', the response includes a `by_venture` map keyed by
-   * venture code. Rows where venture IS NULL are excluded from the map
-   * but still count toward the top-level `total`, so the invariant is
-   * `Σ by_venture[v].total ≤ total`.
-   */
-  group_by?: 'venture'
-}
-
-export interface NotificationCountsResult {
-  total: number
-  by_severity: {
-    critical: number
-    warning: number
-    info: number
-  }
-  by_status: {
-    new: number
-    acked: number
-    resolved: number
-  }
-  by_venture?: Record<string, { critical: number; warning: number; info: number; total: number }>
-  window: {
-    retention_days: number
-    filters: CountNotificationsParams
-  }
-}
-
-export async function countNotifications(
-  db: D1Database,
-  params: CountNotificationsParams
-): Promise<NotificationCountsResult> {
-  const retentionCutoff = new Date(
-    Date.now() - NOTIFICATION_RETENTION_DAYS * 24 * 60 * 60 * 1000
-  ).toISOString()
-
-  const conditions: string[] = ['created_at > ?']
-  const binds: (string | number)[] = [retentionCutoff]
-
-  if (params.status) {
-    conditions.push('status = ?')
-    binds.push(params.status)
-  }
-  if (params.severity) {
-    conditions.push('severity = ?')
-    binds.push(params.severity)
-  }
-  if (params.venture) {
-    conditions.push('venture = ?')
-    binds.push(params.venture)
-  }
-  if (params.repo) {
-    conditions.push('repo = ?')
-    binds.push(params.repo)
-  }
-  if (params.source) {
-    conditions.push('source = ?')
-    binds.push(params.source)
-  }
-
-  const where = `WHERE ${conditions.join(' AND ')}`
-
-  // Single query that produces both severity and status breakdowns via
-  // GROUP BY. Two grouping queries would be cleaner but D1 round-trips
-  // are expensive; one query with both groupings is faster.
-  const severitySql = `SELECT severity, COUNT(*) as count FROM notifications ${where} GROUP BY severity`
-  const statusSql = `SELECT status, COUNT(*) as count FROM notifications ${where} GROUP BY status`
-  // Per-venture breakdown is opt-in via group_by='venture'. Excludes
-  // rows where venture IS NULL (legacy/seed) — they still count toward
-  // the top-level `total` from severitySql, but can't be attributed.
-  const ventureSql = `SELECT venture, severity, COUNT(*) as count FROM notifications ${where} AND venture IS NOT NULL GROUP BY venture, severity`
-
-  const wantVenture = params.group_by === 'venture'
-
-  const [sevResult, statusResult, ventureResult] = await Promise.all([
-    db
-      .prepare(severitySql)
-      .bind(...binds)
-      .all<{ severity: string; count: number }>(),
-    db
-      .prepare(statusSql)
-      .bind(...binds)
-      .all<{ status: string; count: number }>(),
-    wantVenture
-      ? db
-          .prepare(ventureSql)
-          .bind(...binds)
-          .all<{ venture: string; severity: string; count: number }>()
-      : Promise.resolve(null),
-  ])
-
-  const by_severity = { critical: 0, warning: 0, info: 0 }
-  let total = 0
-  for (const row of sevResult.results || []) {
-    total += row.count
-    if (row.severity === 'critical') by_severity.critical = row.count
-    if (row.severity === 'warning') by_severity.warning = row.count
-    if (row.severity === 'info') by_severity.info = row.count
-  }
-
-  const by_status = { new: 0, acked: 0, resolved: 0 }
-  for (const row of statusResult.results || []) {
-    if (row.status === 'new') by_status.new = row.count
-    if (row.status === 'acked') by_status.acked = row.count
-    if (row.status === 'resolved') by_status.resolved = row.count
-  }
-
-  let by_venture:
-    | Record<string, { critical: number; warning: number; info: number; total: number }>
-    | undefined
-  if (ventureResult) {
-    by_venture = {}
-    for (const row of ventureResult.results || []) {
-      const v = row.venture
-      const bucket = by_venture[v] ?? { critical: 0, warning: 0, info: 0, total: 0 }
-      if (row.severity === 'critical') bucket.critical = row.count
-      if (row.severity === 'warning') bucket.warning = row.count
-      if (row.severity === 'info') bucket.info = row.count
-      bucket.total = bucket.critical + bucket.warning + bucket.info
-      by_venture[v] = bucket
+// buildResolveQuery is declared after processGreenEvent (function hoisting makes it
+// callable above). Both branches embed the `auto_resolved_by_id IS NULL` predicate
+// so the SQL-invariant test can find at least two occurrences in the post-declaration
+// source slice.
+function buildResolveQuery(
+  greenId: string,
+  now: string,
+  cutoff: string,
+  p: ProcessGreenEventParams
+): { sql: string; binds: (string | number | null)[] } {
+  if (p.is_schedule_like) {
+    return {
+      sql: `UPDATE notifications
+       SET status = 'resolved', auto_resolved_by_id = ?, auto_resolve_reason = ?,
+           resolved_at = ?, updated_at = ?
+       WHERE match_key = ? AND status IN ('new', 'acked')
+         AND auto_resolved_by_id IS NULL AND created_at > ?
+         AND head_sha = ?
+         AND (run_started_at IS NULL OR run_started_at <= ?)`,
+      binds: [
+        greenId,
+        p.auto_resolve_reason,
+        now,
+        now,
+        p.match_key,
+        cutoff,
+        p.head_sha,
+        p.run_started_at,
+      ],
     }
   }
-
   return {
-    total,
-    by_severity,
-    by_status,
-    ...(by_venture !== undefined ? { by_venture } : {}),
-    window: {
-      retention_days: NOTIFICATION_RETENTION_DAYS,
-      filters: params,
-    },
+    sql: `UPDATE notifications
+       SET status = 'resolved', auto_resolved_by_id = ?, auto_resolve_reason = ?,
+           resolved_at = ?, updated_at = ?
+       WHERE match_key = ? AND status IN ('new', 'acked')
+         AND auto_resolved_by_id IS NULL AND created_at > ?
+         AND (run_started_at IS NULL OR run_started_at <= ?)`,
+    binds: [greenId, p.auto_resolve_reason, now, now, p.match_key, cutoff, p.run_started_at],
   }
-}
-
-/**
- * Get the oldest notification matching a filter. Used by the
- * notification-retention-window health check (plan §B.7) to verify the
- * retention filter is actually working: if the oldest open notification
- * is older than NOTIFICATION_RETENTION_DAYS, something is broken.
- */
-export async function getOldestNotification(
-  db: D1Database,
-  params: CountNotificationsParams
-): Promise<NotificationRecord | null> {
-  const conditions: string[] = []
-  const binds: (string | number)[] = []
-
-  if (params.status) {
-    conditions.push('status = ?')
-    binds.push(params.status)
-  }
-  if (params.venture) {
-    conditions.push('venture = ?')
-    binds.push(params.venture)
-  }
-  if (params.severity) {
-    conditions.push('severity = ?')
-    binds.push(params.severity)
-  }
-
-  const where = conditions.length > 0 ? `WHERE ${conditions.join(' AND ')}` : ''
-  const sql = `SELECT * FROM notifications ${where} ORDER BY created_at ASC LIMIT 1`
-
-  return await db
-    .prepare(sql)
-    .bind(...binds)
-    .first<NotificationRecord>()
 }
